Splitting Extract text into separate file
